reuse amb choices from execution context
During evaluation, we keep track of any values we've chosen
from various amb literals, and we reuse those same choices
whenever we perform a cell reference that depends on the same
amb literals.

The result is that the dimensionality of any formula result
is equal to the number of amb literals in its transitive dependencies.
(And the total number of distinct values in the formula result
is equal to the product of the number of values in the
amb literals.)

This is implemented by passing around a context
object in the evaluator which stores a mapping from amb nodes
to a numeric index tracking which value was chosen for that node.