remove codemirror selection drawing for now
codemirror has the ability to draw a selection instead of relying
on the browser, which helps for preserving the selection after
the document is blurred. but this feature wasn't behaving well
in the presence of wrapped lines. so for now we're disabling it.

the downside is that we now don't preserve selection while you are
leaving a comment on selected text. we'll need to find a better
solution to bring that back.